refactoring changes after rebasing to main
